requirements.txt
Django==1.7.1
argparse==1.2.1
django-phonenumber-field==0.7.2
phonenumbers==7.0.4
